+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 ## v0.1.3
 ### Added
 - Added context variable for the current page URL.
+
+### Fixes
+- Fixed issue where blank page default was not populated.